feat: 1c transfer create fixed and cart_item fixed sale type online then checked, add filter transfer detail expected_count

+++ internal/controller/http/v1/cart_item.handler.go
@@ -69,15 +69,15 @@ func (h *CartItemHandler) Create(c *gin.Context) {
 	ctx, cancel := context.WithTimeout(c.Request.Context(), constants.DefaultContextTimeout)
 	defer cancel()
 
-	// var sale domain.Sale
-	// if err := h.db.WithContext(ctx).Select("type").Where("id = ?", body.SaleId).First(&sale).Error; err != nil {
-	// 	handleServiceResponse(c, InternalError, err)
-	// 	return
-	// }
-	// if sale.Type == "online" {
-	// 	handleServiceResponse(c, FORBIDDEN, domain.ForbiddinError)
-	// 	return
-	// }
+	var sale domain.Sale
+	if err := h.db.WithContext(ctx).Select("type").Where("id = ?", body.SaleId).First(&sale).Error; err != nil {
+		handleServiceResponse(c, InternalError, err)
+		return
+	}
+	if sale.Type == "online" {
+		handleServiceResponse(c, BadRequest, domain.BadRequestError)
+		return
+	}
 
 	body.EmployeeId = user.UserId
 
@@ -400,15 +400,15 @@ func (h *CartItemHandler) Delete(c *gin.Context) {
 		return
 	}
 
-	// var sale domain.Sale
-	// if err := h.db.WithContext(ctx).Select("type").Where("id = ?", cartItem.SaleId).First(&sale).Error; err != nil {
-	// 	handleServiceResponse(c, InternalError, err)
-	// 	return
-	// }
-	// if sale.Type == "online" {
-	// 	handleServiceResponse(c, FORBIDDEN, domain.ForbiddinError)
-	// 	return
-	// }
+	var sale domain.Sale
+	if err := h.db.WithContext(ctx).Select("type").Where("id = ?", cartItem.SaleId).First(&sale).Error; err != nil {
+		handleServiceResponse(c, InternalError, err)
+		return
+	}
+	if sale.Type == "online" {
+		handleServiceResponse(c, BadRequest, domain.BadRequestError)
+		return
+	}
 
 	err := h.service.DeleteCartItem(ctx, id)
 	if err != nil {
